Explicitly enumerate the whitespace characters.
I don't want this library to depend on Regex anymore. Once an Email
validation Parser exists, I want to switch to that!

In preparation for that, I don't want the whitespace checker to depend
on Regex anymore either. Instead, this has it explicitly enumerate in
the docs what characters it considers whitespace, and check for them
using String.uncons.